move backtest and timeseries metrics to pyseir folder
